moved executionContext in its own folder. Move loaders in game folder
